updating per comments
