+++ src/core/imageUtils.js
@@ -6,3 +6,14 @@ export function getAddonIconUrl(addon) {
     ? addon.icon_url
     : fallbackIcon;
 }
+
+export const getPreviewImage = (addon, { index = 0, full = true } = {}) => {
+  const preview = addon.previews.length && addon.previews[index];
+  if (preview) {
+    const previewSize = full ? 'image_url' : 'thumbnail_url';
+    return preview[previewSize] && isAllowedOrigin(preview[previewSize])
+      ? preview[previewSize]
+      : null;
+  }
+  return null;
+};
