+++ src/core/utils/index.js
@@ -315,13 +315,3 @@ export const addQueryParamsToHistory = ({
   return qhistory(history, _stringify, _parse);
 };
 
-export const getPreviewImage = (addon, { version = 0, full = true } = {}) => {
-  const preview = addon.previews.length && addon.previews[version];
-  if (preview) {
-    const previewSize = full ? 'image_url' : 'thumbnail_url';
-    return preview[previewSize] && isAllowedOrigin(preview[previewSize])
-      ? preview[previewSize]
-      : null;
-  }
-  return null;
-};
