+++ tests/unit/core/utils/test_index.js
@@ -28,7 +28,6 @@ import {
   getCategoryColor,
   getClientApp,
   getClientConfig,
-  getPreviewImage,
   isAddonAuthor,
   isAllowedOrigin,
   isTheme,
@@ -45,11 +44,7 @@ import {
   trimAndAddProtocolToUrl,
 } from 'core/utils';
 import { createInternalAddon, loadAddons } from 'core/reducers/addons';
-import {
-  fakeAddon,
-  fakePreview,
-  signedInApiState,
-} from 'tests/unit/amo/helpers';
+import { fakeAddon, signedInApiState } from 'tests/unit/amo/helpers';
 import {
   createFakeHistory,
   createFakeLocation,
@@ -795,83 +790,6 @@ describe(__filename, () => {
       expect(historyWithQueryParams).toHaveProperty('location.query', {
         foo: '123',
       });
-    });
-  });
-
-  describe('getPreviewImage', () => {
-    it('returns the first full image from the previews array', () => {
-      const fullImage = 'https://addons.cdn.mozilla.net/full/12345.png';
-      const addon = createInternalAddon({
-        previews: [
-          {
-            ...fakePreview,
-            image_url: fullImage,
-          },
-          {
-            ...fakePreview,
-            image_url: 'https://addons.cdn.mozilla.net/image.not.used.here.png',
-          },
-        ],
-      });
-      const image = getPreviewImage(addon);
-      expect(image).toEqual(fullImage);
-    });
-
-    it('returns the full image from the 2nd item in the previews array', () => {
-      const fullImage = 'https://addons.cdn.mozilla.net/full/12345.png';
-      const addon = createInternalAddon({
-        previews: [
-          {
-            ...fakePreview,
-            image_url: 'https://addons.cdn.mozilla.net/image.not.used.here.png',
-          },
-          {
-            ...fakePreview,
-            image_url: fullImage,
-          },
-        ],
-      });
-
-      const image = getPreviewImage(addon, { version: 1 });
-      expect(image).toEqual(fullImage);
-    });
-
-    it('returns the thumb image from the previews array', () => {
-      const thumbImage = 'https://addons.cdn.mozilla.net/full/12345.png';
-      const addon = createInternalAddon({
-        previews: [
-          {
-            ...fakePreview,
-            thumbnail_url: thumbImage,
-          },
-        ],
-      });
-
-      const image = getPreviewImage(addon, { full: false });
-      expect(image).toEqual(thumbImage);
-    });
-
-    it('returns null if the previews array is empty', () => {
-      const addon = createInternalAddon({
-        previews: [],
-      });
-
-      const image = getPreviewImage(addon);
-      expect(image).toEqual(null);
-    });
-
-    it('returns null if the isAllowedOrigin returns false', () => {
-      const addon = createInternalAddon({
-        previews: [
-          {
-            ...fakePreview,
-            image_url: 'http://www.example.com',
-          },
-        ],
-      });
-
-      const image = getPreviewImage(addon);
-      expect(image).toEqual(null);
     });
   });
 });
